Create continueStatement.js
The continue statement tells the interpreter to immediately start the next iteration of the loop and skip the remaining code block. When
a continue statement is encountered, the program flow moves to the loop check expression immediately and if the condition remains true, then it starts the next iteration, otherwise the control comes out of the loop.

This example illustrates the use of a continue statement with a while loop. Notice how the continue statement is used to skip printing when the index held in variable x reaches 5.